Move VAOS unit tests: Batch 4
