+++ src/applications/vaos/components/AppointmentCard/index.unit.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { expect } from 'chai';
 import { renderWithStoreAndRouter } from '~/platform/testing/unit/react-testing-library-helpers';
-import AppointmentCard from '../AppointmentCard';
+import AppointmentCard from '.';
 import AppointmentDateTime from '../../appointment-list/components/AppointmentDateTime';
 import { createTestStore } from '../../tests/mocks/setup';
