+++ src/applications/vaos/components/layout/CCRequestLayout.unit.spec.js
@@ -4,7 +4,7 @@ import {
   createTestStore,
   renderWithStoreAndRouter,
 } from '../../tests/mocks/setup';
-import VARequestLayout from '../layout/VARequestLayout';
+import CCRequestLayout from './CCRequestLayout';
 
 describe('VAOS Component: VARequestLayout', () => {
   const initialState = {
@@ -32,13 +32,12 @@ describe('VAOS Component: VARequestLayout', () => {
     },
   };
 
-  describe('When viewing request details page', () => {
-    it('should display VA request layout', async () => {
+  describe('When viewing upcomming CC appointment details', () => {
+    it('should display CC request layout', async () => {
       // Arrange
       const store = createTestStore(initialState);
       const appointment = {
-        reasonForAppointment: 'This is a test',
-        patientComments: 'Additional information:colon',
+        patientComments: 'This is a test:Additional information',
         contact: {
           telecom: [
             {
@@ -59,7 +58,7 @@ describe('VAOS Component: VARequestLayout', () => {
         preferredDates: [],
         videoData: {},
         vaos: {
-          isCommunityCare: false,
+          isCommunityCare: true,
           isCompAndPenAppointment: false,
           isCOVIDVaccine: false,
           isPendingAppointment: true,
@@ -73,18 +72,20 @@ describe('VAOS Component: VARequestLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <VARequestLayout data={appointment} />,
+        <CCRequestLayout data={appointment} />,
         {
           store,
         },
       );
+
       // Assert
       expect(
         screen.getByRole('heading', {
           level: 1,
-          name: /Request for appointment/i,
+          name: /Request for community care appointment/i,
         }),
       );
+
       expect(
         screen.getByText(
           /We.ll try to schedule your appointment in the next 2 business days/i,
@@ -111,19 +112,24 @@ describe('VAOS Component: VARequestLayout', () => {
       expect(screen.getByRole('heading', { level: 2, name: /Type of care/i }));
       expect(screen.getByText(/Primary care/i));
 
-      expect(screen.getByText(/How you prefer to attend/i));
-
-      expect(screen.getByRole('heading', { level: 2, name: /Facility/i }));
-      expect(screen.getByText(/Cheyenne VA Medical Center/i));
-      expect(screen.getByText(/2360 East Pershing Boulevard/i));
-
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
-        .ok;
-
-      expect(screen.getByRole('heading', { level: 2, name: /Phone/i }));
       expect(
-        screen.container.querySelector('va-telephone[contact="307-778-7550"]'),
-      ).to.be.ok;
+        screen.getByRole('heading', { level: 2, name: /Scheduling facility/i }),
+      );
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Preferred community care provider/i,
+        }),
+      );
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Language you.d prefer the provider speak/i,
+        }),
+      );
+
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
+        .to.exist;
 
       expect(
         screen.getByRole('heading', {
@@ -131,8 +137,11 @@ describe('VAOS Component: VARequestLayout', () => {
           name: /Details you.d like to share with your provider/i,
         }),
       );
-      expect(screen.getByText(/Reason: This is a test/i));
-      expect(screen.getByText(/Other details: Additional information:colon/i));
+      expect(
+        screen.getByText(
+          /Other details: This is a test:Additional information/i,
+        ),
+      );
 
       expect(
         screen.getByRole('heading', {
@@ -157,22 +166,15 @@ describe('VAOS Component: VARequestLayout', () => {
         .ok;
       expect(screen.container.querySelector('va-button[text="Cancel request"]'))
         .to.be.ok;
-      expect(
-        screen.queryByRole('heading', {
-          level: 2,
-          name: /After visit summary/i,
-        }),
-      ).to.be.null;
     });
   });
 
-  describe('When viewing canceled appointment details', () => {
-    it('should display VA request layout', async () => {
+  describe('When scheduling an CC appointment request', () => {
+    it('should display CC request layout', async () => {
       // Arrange
       const store = createTestStore(initialState);
       const appointment = {
-        reasonForAppointment: 'This is a test',
-        patientComments: 'Additional information:colon',
+        patientComments: 'This is a test:Additional information',
         contact: {
           telecom: [
             {
@@ -193,126 +195,7 @@ describe('VAOS Component: VARequestLayout', () => {
         preferredDates: [],
         videoData: {},
         vaos: {
-          isCommunityCare: false,
-          isCompAndPenAppointment: false,
-          isCOVIDVaccine: false,
-          isPendingAppointment: true,
-          isUpcomingAppointment: false,
-          apiData: {
-            serviceType: 'primaryCare',
-          },
-        },
-        status: 'cancelled',
-      };
-
-      // Act
-      const screen = renderWithStoreAndRouter(
-        <VARequestLayout data={appointment} />,
-        {
-          store,
-        },
-      );
-
-      // Assert
-      expect(
-        screen.getByRole('heading', {
-          level: 1,
-          name: /Canceled request for appointment/i,
-        }),
-      );
-      expect(
-        screen.container.querySelector(
-          'va-link[text="Schedule a new appointment"]',
-        ),
-      ).to.be.ok;
-      expect(screen.getByText(/Facility canceled this appointment/i));
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Preferred date and time/i,
-        }),
-      );
-
-      expect(screen.getByRole('heading', { level: 2, name: /Type of care/i }));
-      expect(screen.getByText(/Primary care/i));
-
-      expect(screen.getByText(/How you prefer to attend/i));
-
-      expect(screen.getByRole('heading', { level: 2, name: /Facility/i }));
-      expect(screen.getByText(/Cheyenne VA Medical Center/i));
-      expect(screen.getByText(/2360 East Pershing Boulevard/i));
-
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
-        .ok;
-
-      expect(screen.getByRole('heading', { level: 2, name: /Phone/i }));
-      expect(
-        screen.container.querySelector('va-telephone[contact="307-778-7550"]'),
-      ).to.be.ok;
-
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Details you.d like to share with your provider/i,
-        }),
-      );
-      expect(screen.getByText(/Reason: This is a test/i));
-      expect(screen.getByText(/Other details: Additional information:colon/i));
-
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Your contact details/i,
-        }),
-      );
-      expect(
-        screen.getByText((content, element) => {
-          return (
-            element.tagName.toLowerCase() === 'span' &&
-            content === 'Email: user@va.gov'
-          );
-        }),
-      );
-      expect(screen.getByText(/Phone number/));
-      expect(
-        screen.container.querySelector('va-telephone[contact="1234567890"]'),
-      ).to.be.ok;
-
-      expect(screen.container.querySelector('va-button[text="Print"]')).to.be
-        .ok;
-      expect(screen.container.querySelector('va-button[text="Cancel request"]'))
-        .not.to.exist;
-    });
-  });
-
-  describe('When scheduling an appointment request', () => {
-    it('should display VA request layout', async () => {
-      // Arrange
-      const store = createTestStore(initialState);
-      const appointment = {
-        reasonForAppointment: 'This is a test',
-        patientComments: 'Additional information:colon',
-        contact: {
-          telecom: [
-            {
-              type: 'email',
-              value: 'user@va.gov',
-            },
-            {
-              type: 'phone',
-              value: '1234567890',
-            },
-          ],
-        },
-        location: {
-          stationId: '983',
-          clinicName: 'Clinic 1',
-          clinicPhysicalLocation: 'CHEYENNE',
-        },
-        preferredDates: [],
-        videoData: {},
-        vaos: {
-          isCommunityCare: false,
+          isCommunityCare: true,
           isCompAndPenAppointment: false,
           isCOVIDVaccine: false,
           isPendingAppointment: true,
@@ -326,7 +209,7 @@ describe('VAOS Component: VARequestLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <VARequestLayout data={appointment} />,
+        <CCRequestLayout data={appointment} />,
         {
           store,
           path: '/?confirmMsg=true',
@@ -340,6 +223,7 @@ describe('VAOS Component: VARequestLayout', () => {
           name: /We have received your request/i,
         }),
       );
+
       expect(
         screen.getByText(
           /We.ll try to schedule your appointment in the next 2 business days/i,
@@ -366,19 +250,24 @@ describe('VAOS Component: VARequestLayout', () => {
       expect(screen.getByRole('heading', { level: 2, name: /Type of care/i }));
       expect(screen.getByText(/Primary care/i));
 
-      expect(screen.getByText(/How you prefer to attend/i));
-
-      expect(screen.getByRole('heading', { level: 2, name: /Facility/i }));
-      expect(screen.getByText(/Cheyenne VA Medical Center/i));
-      expect(screen.getByText(/2360 East Pershing Boulevard/i));
-
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
-        .ok;
-
-      expect(screen.getByRole('heading', { level: 2, name: /Phone/i }));
       expect(
-        screen.container.querySelector('va-telephone[contact="307-778-7550"]'),
-      ).to.be.ok;
+        screen.getByRole('heading', { level: 2, name: /Scheduling facility/i }),
+      );
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Preferred community care provider/i,
+        }),
+      );
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Language you.d prefer the provider speak/i,
+        }),
+      );
+
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
+        .to.exist;
 
       expect(
         screen.getByRole('heading', {
@@ -386,8 +275,11 @@ describe('VAOS Component: VARequestLayout', () => {
           name: /Details you.d like to share with your provider/i,
         }),
       );
-      expect(screen.getByText(/Reason: This is a test/i));
-      expect(screen.getByText(/Other details: Additional information:colon/i));
+      expect(
+        screen.getByText(
+          /Other details: This is a test:Additional information/i,
+        ),
+      );
 
       expect(
         screen.getByRole('heading', {
