+++ src/applications/vaos/components/layout/InPersonLayout.unit.spec.js
@@ -5,9 +5,9 @@ import {
   createTestStore,
   renderWithStoreAndRouter,
 } from '../../tests/mocks/setup';
-import PhoneLayout from '../layout/PhoneLayout';
+import InPersonLayout from './InPersonLayout';
 
-describe('VAOS Component: PhoneLayout', () => {
+describe('VAOS Component: InPersonLayout', () => {
   const initialState = {
     appointments: {
       facilityData: {
@@ -31,6 +31,101 @@ describe('VAOS Component: PhoneLayout', () => {
   };
 
   describe('When appointment information is missing', () => {
+    it('should display view facility info when only facility id is returned', async () => {
+      // Arrange
+      const state = {
+        ...initialState,
+        appointments: {
+          facilityData: {},
+        },
+      };
+      const store = createTestStore(state);
+
+      const appointment = {
+        location: {
+          vistaId: '983',
+          clinicId: '848',
+          stationId: '983',
+          clinicName: 'CHY PC VAR2',
+        },
+        videoData: {},
+        vaos: {
+          isCommunityCare: false,
+          isCompAndPenAppointment: false,
+          isCOVIDVaccine: false,
+          isPendingAppointment: false,
+          isUpcomingAppointment: true,
+        },
+        status: 'booked',
+      };
+
+      // Act
+      const screen = renderWithStoreAndRouter(
+        <InPersonLayout data={appointment} />,
+        {
+          store,
+        },
+      );
+
+      // Assert
+      expect(
+        screen.getByRole('heading', { level: 2, name: /Where to attend/i }),
+      );
+      expect(screen.getByText(/Facility details not available/i));
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
+        .to.exist;
+      expect(
+        screen.getByRole('link', {
+          name: /View facility information Link opens in a new tab./i,
+        }),
+      );
+    });
+
+    it('should display find facility info when no facility data and no facility id are available', async () => {
+      // Arrange
+      const state = {
+        ...initialState,
+        appointments: {
+          facilityData: {},
+        },
+      };
+      const store = createTestStore(state);
+
+      const appointment = {
+        location: {},
+        videoData: {},
+        vaos: {
+          isCommunityCare: false,
+          isCompAndPenAppointment: false,
+          isCOVIDVaccine: false,
+          isPendingAppointment: false,
+          isUpcomingAppointment: true,
+        },
+        status: 'booked',
+      };
+
+      // Act
+      const screen = renderWithStoreAndRouter(
+        <InPersonLayout data={appointment} />,
+        {
+          store,
+        },
+      );
+
+      // Assert
+      expect(
+        screen.getByRole('heading', { level: 2, name: /Where to attend/i }),
+      );
+      expect(screen.getByText(/Facility details not available/i));
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
+        .to.exist;
+      expect(
+        screen.getByRole('link', {
+          name: /Find facility information Link opens in a new tab/i,
+        }),
+      );
+    });
+
     it('should not display heading and text for empty data', async () => {
       // Arrange
       const store = createTestStore(initialState);
@@ -39,25 +134,12 @@ describe('VAOS Component: PhoneLayout', () => {
           stationId: '983',
         },
         videoData: {},
-        practitioners: [
-          {
-            identifier: [
-              {
-                system: 'dfn-983',
-                value: '520647363',
-              },
-            ],
-            practiceName: 'Cheyenne VA Medical Center',
-          },
-        ],
         vaos: {
           isCommunityCare: false,
           isCompAndPenAppointment: false,
           isCOVIDVaccine: false,
           isPendingAppointment: false,
           isUpcomingAppointment: true,
-          isPhoneAppointment: true,
-          isCancellable: true,
           apiData: {},
         },
         status: 'booked',
@@ -65,44 +147,49 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
       );
 
       // Assert
+      expect(screen.queryByRole('heading', { level: 2, name: /What/i })).not.to
+        .exist;
+      expect(screen.queryByRole('heading', { level: 2, name: /Who/i })).not.to
+        .exist;
       expect(
-        screen.getByRole('heading', {
-          level: 1,
-          name: /Phone appointment/i,
+        screen.getByText((content, element) => {
+          return (
+            element.tagName.toLowerCase() === 'span' &&
+            content === 'Location: Not available'
+          );
         }),
       );
-
       expect(
-        screen.queryByRole('heading', {
-          level: 2,
-          name: /Who/i,
-        }),
-      ).not.to.exist;
-
-      expect(
-        screen.queryByRole('heading', {
-          level: 2,
-          name: /What/i,
-        }),
-      ).not.to.exist;
-
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Details you shared with your provider/i,
+        screen.getByText((content, element) => {
+          return (
+            element.tagName.toLowerCase() === 'span' &&
+            content === 'Clinic: Not available'
+          );
         }),
       );
-
-      expect(screen.getByText(/Clinic: Not available/i));
-      expect(screen.getByText(/Reason: Not available/i));
-      expect(screen.getByText(/Other details: Not available/i));
+      expect(
+        screen.getByText((content, element) => {
+          return (
+            element.tagName.toLowerCase() === 'span' &&
+            content === 'Reason: Not available'
+          );
+        }),
+      );
+      expect(
+        screen.getByText((content, element) => {
+          return (
+            element.tagName.toLowerCase() === 'span' &&
+            content === 'Other details: Not available'
+          );
+        }),
+      );
     });
     it('should display facility phone when clinic phone is missing', async () => {
       // Arrange
@@ -118,8 +205,6 @@ describe('VAOS Component: PhoneLayout', () => {
           isCOVIDVaccine: false,
           isPendingAppointment: false,
           isUpcomingAppointment: true,
-          isPhoneAppointment: true,
-          isCancellable: true,
           apiData: {},
         },
         status: 'booked',
@@ -127,7 +212,7 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
@@ -137,8 +222,7 @@ describe('VAOS Component: PhoneLayout', () => {
         screen.container.querySelector('va-telephone[contact="307-778-7550"]'),
       ).to.be.ok;
     });
-
-    it('should display VA main phone when facility id is missing', async () => {
+    it('should display display VA main phone when facility id is missing', async () => {
       // Arrange
       const store = createTestStore(initialState);
       const appointment = {
@@ -150,8 +234,6 @@ describe('VAOS Component: PhoneLayout', () => {
           isCOVIDVaccine: false,
           isPendingAppointment: false,
           isUpcomingAppointment: true,
-          isPhoneAppointment: true,
-          isCancellable: true,
           apiData: {},
         },
         status: 'booked',
@@ -159,7 +241,7 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
@@ -172,19 +254,12 @@ describe('VAOS Component: PhoneLayout', () => {
   });
 
   describe('When viewing upcoming appointment details', () => {
-    it('should display phone layout', async () => {
+    it('should display in-person layout', async () => {
       // Arrange
       const store = createTestStore(initialState);
       const appointment = {
         reasonForAppointment: 'This is a test',
         patientComments: 'Additional information:colon',
-        location: {
-          stationId: '983',
-          clinicName: 'Clinic 1',
-          clinicPhysicalLocation: 'CHEYENNE',
-          clinicPhone: '500-500-5000',
-          clinicPhoneExtension: '1234',
-        },
         practitioners: [
           {
             name: {
@@ -193,6 +268,13 @@ describe('VAOS Component: PhoneLayout', () => {
             },
           },
         ],
+        location: {
+          stationId: '983',
+          clinicName: 'Clinic 1',
+          clinicPhysicalLocation: 'CHEYENNE',
+          clinicPhone: '500-500-5000',
+          clinicPhoneExtension: '1234',
+        },
         videoData: {},
         vaos: {
           isCommunityCare: false,
@@ -200,7 +282,6 @@ describe('VAOS Component: PhoneLayout', () => {
           isCOVIDVaccine: false,
           isPendingAppointment: false,
           isUpcomingAppointment: true,
-          isPhoneAppointment: true,
           isCancellable: true,
           apiData: {
             serviceType: 'primaryCare',
@@ -211,7 +292,7 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
@@ -221,10 +302,9 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.getByRole('heading', {
           level: 1,
-          name: /Phone appointment/i,
+          name: /In-person appointment/i,
         }),
       );
-      expect(screen.getByRole('heading', { level: 2, name: /How to join/i }));
       expect(screen.getByRole('heading', { level: 2, name: /When/i }));
       expect(
         screen.container.querySelector('va-button[text="Add to calendar"]'),
@@ -237,15 +317,16 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(screen.getByText(/Test User/i));
 
       expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Scheduling facility/i,
-        }),
+        screen.getByRole('heading', { level: 2, name: /Where to attend/i }),
       );
       expect(screen.getByText(/Cheyenne VA Medical Center/i));
       expect(screen.getByText(/2360 East Pershing Boulevard/i));
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
-        .to.exist;
+
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
+        .ok;
+
+      expect(screen.getByText(/Location:/i));
+      expect(screen.getByText(/CHEYENNE/));
 
       expect(screen.getByText(/Clinic:/i));
       expect(screen.getByText(/Clinic 1/i));
@@ -259,15 +340,6 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.getByRole('heading', {
           level: 2,
-          name: /Details you shared with your provider/i,
-        }),
-      );
-      expect(screen.getByText(/Reason: This is a test/i));
-      expect(screen.getByText(/Other details: Additional information:colon/i));
-
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
           name: /Prepare for your appointment/i,
         }),
       );
@@ -287,12 +359,22 @@ describe('VAOS Component: PhoneLayout', () => {
         ),
       ).to.be.ok;
 
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Details you shared with your provider/i,
+        }),
+      );
+      expect(screen.getByText(/Reason: This is a test/i));
+      expect(screen.getByText(/Other details: Additional information:colon/i));
+
       expect(screen.container.querySelector('va-button[text="Print"]'));
       expect(
         screen.container.querySelector('va-button[text="Cancel appointment"]'),
       ).to.be.ok;
     });
-    it('should display phone layout without cancel button', async () => {
+
+    it('should display in-person layout without cancel button', async () => {
       // Arrange
       const store = createTestStore(initialState);
       const appointment = {
@@ -310,7 +392,6 @@ describe('VAOS Component: PhoneLayout', () => {
           isCOVIDVaccine: false,
           isPendingAppointment: false,
           isUpcomingAppointment: true,
-          isPhoneAppointment: true,
           isCancellable: false,
           apiData: {
             serviceType: 'primaryCare',
@@ -321,7 +402,7 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
@@ -331,10 +412,9 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.getByRole('heading', {
           level: 1,
-          name: /Phone appointment/i,
+          name: /In-person appointment/i,
         }),
       );
-      expect(screen.getByRole('heading', { level: 2, name: /How to join/i }));
       expect(screen.getByRole('heading', { level: 2, name: /When/i }));
       expect(
         screen.container.querySelector('va-button[text="Add to calendar"]'),
@@ -344,15 +424,16 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(screen.getByText(/Primary care/i));
 
       expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Scheduling facility/i,
-        }),
+        screen.getByRole('heading', { level: 2, name: /Where to attend/i }),
       );
       expect(screen.getByText(/Cheyenne VA Medical Center/i));
       expect(screen.getByText(/2360 East Pershing Boulevard/i));
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
-        .to.exist;
+
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
+        .ok;
+
+      expect(screen.getByText(/Location:/i));
+      expect(screen.getByText(/CHEYENNE/));
 
       expect(screen.getByText(/Clinic:/i));
       expect(screen.getByText(/Clinic 1/i));
@@ -360,15 +441,6 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.container.querySelector('va-telephone[contact="307-778-7550"]'),
       ).to.be.ok;
-
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Details you shared with your provider/i,
-        }),
-      );
-      expect(screen.getByText(/Reason: This is a test/i));
-      expect(screen.getByText(/Other details: Additional information:colon/i));
 
       expect(
         screen.getByRole('heading', {
@@ -392,15 +464,24 @@ describe('VAOS Component: PhoneLayout', () => {
         ),
       ).to.be.ok;
 
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Details you shared with your provider/i,
+        }),
+      );
+      expect(screen.getByText(/Reason: This is a test/i));
+      expect(screen.getByText(/Other details: Additional information:colon/i));
+
       expect(screen.container.querySelector('va-button[text="Print"]'));
       expect(
         screen.container.querySelector('va-button[text="Cancel appointment"]'),
-      ).to.not.exist;
+      ).not.to.exist;
     });
   });
 
   describe('When viewing past appointment details', () => {
-    it('should display phone layout', async () => {
+    it('should display in-person layout', async () => {
       // Arrange
       const store = createTestStore(initialState);
       const appointment = {
@@ -429,7 +510,7 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
@@ -439,11 +520,9 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.getByRole('heading', {
           level: 1,
-          name: /Past phone appointment/i,
+          name: /Past in-person appointment/i,
         }),
       );
-      expect(screen.queryByRole('heading', { level: 2, name: /How to join/ }))
-        .not.to.exist;
       expect(
         screen.getByRole('heading', { level: 2, name: /After visit summary/i }),
       );
@@ -456,13 +535,13 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(screen.getByRole('heading', { level: 2, name: /What/i }));
       expect(screen.getByText(/Primary care/i));
 
-      expect(
-        screen.getByRole('heading', { level: 2, name: /Scheduling facility/i }),
-      );
-      expect(screen.getByText(/Cheyenne VA Medical Center/i));
+      expect(screen.getByRole('heading', { level: 2, name: /Where/i }));
       expect(screen.getByText(/2360 East Pershing Boulevard/i));
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
-        .to.exist;
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
+        .ok;
+
+      expect(screen.getByText(/Location:/i));
+      expect(screen.getByText(/CHEYENNE/));
 
       expect(screen.getByText(/Clinic:/i));
       expect(screen.getByText(/Clinic 1/i));
@@ -472,6 +551,12 @@ describe('VAOS Component: PhoneLayout', () => {
       ).to.be.ok;
 
       expect(
+        screen.queryByRole('heading', {
+          name: /Prepare for your appointment/i,
+        }),
+      ).not.to.exist;
+
+      expect(
         screen.getByRole('heading', {
           level: 2,
           name: /Details you shared with your provider/i,
@@ -479,12 +564,6 @@ describe('VAOS Component: PhoneLayout', () => {
       );
       expect(screen.getByText(/Reason: This is a test/i));
       expect(screen.getByText(/Other details: Additional information:colon/i));
-
-      expect(
-        screen.queryByRole('heading', {
-          name: /Prepare for your appointment/i,
-        }),
-      ).not.to.exist;
 
       expect(screen.container.querySelector('va-button[text="Print"]')).to.be
         .ok;
@@ -521,7 +600,7 @@ describe('VAOS Component: PhoneLayout', () => {
 
       // Act
       const screen = renderWithStoreAndRouter(
-        <PhoneLayout data={appointment} />,
+        <InPersonLayout data={appointment} />,
         {
           store,
         },
@@ -531,7 +610,7 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.getByRole('heading', {
           level: 1,
-          name: /Canceled phone appointment/i,
+          name: /Canceled in-person appointment/i,
         }),
       );
       expect(
@@ -539,8 +618,6 @@ describe('VAOS Component: PhoneLayout', () => {
           /If you want to reschedule, call us or schedule a new appointment online/i,
         ),
       );
-      expect(screen.queryByRole('heading', { level: 2, name: /How to join/ }))
-        .not.to.exist;
       expect(
         screen.queryByRole('heading', {
           level: 2,
@@ -556,13 +633,13 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(screen.getByRole('heading', { level: 2, name: /What/i }));
       expect(screen.getByText(/Primary care/i));
 
-      expect(
-        screen.getByRole('heading', { level: 2, name: /Scheduling facility/i }),
-      );
-      expect(screen.getByText(/Cheyenne VA Medical Center/i));
+      expect(screen.getByRole('heading', { level: 2, name: /Where/i }));
       expect(screen.getByText(/2360 East Pershing Boulevard/i));
-      expect(screen.container.querySelector('va-icon[icon="directions"]')).not
-        .to.exist;
+      expect(screen.container.querySelector('va-icon[icon="directions"]')).to.be
+        .ok;
+
+      expect(screen.getByText(/Location:/i));
+      expect(screen.getByText(/CHEYENNE/));
 
       expect(screen.getByText(/Clinic:/i));
       expect(screen.getByText(/Clinic 1/i));
@@ -570,15 +647,6 @@ describe('VAOS Component: PhoneLayout', () => {
       expect(
         screen.container.querySelector('va-telephone[contact="500-500-5000"]'),
       ).to.be.ok;
-
-      expect(
-        screen.getByRole('heading', {
-          level: 2,
-          name: /Details you shared with your provider/i,
-        }),
-      );
-      expect(screen.getByText(/Reason: This is a test/i));
-      expect(screen.getByText(/Other details: Additional information:colon/i));
 
       expect(
         screen.getByRole('heading', {
@@ -601,6 +669,15 @@ describe('VAOS Component: PhoneLayout', () => {
           'va-link[text="Find a full list of things to bring to your appointment"]',
         ),
       ).to.be.ok;
+
+      expect(
+        screen.getByRole('heading', {
+          level: 2,
+          name: /Details you shared with your provider/i,
+        }),
+      );
+      expect(screen.getByText(/Reason: This is a test/i));
+      expect(screen.getByText(/Other details: Additional information:colon/i));
 
       expect(screen.container.querySelector('va-button[text="Print"]')).to.be
         .ok;
